requirements.txt
paho-mqtt
twisted
adafruit-blinka
adafruit-circuitpython-icm20x
termcolor
numpy
sparkfun-qwiic-titan-gps
pygeohash
psutil
adafruit-circuitpython-ads1x15
adafruit-circuitpython-bme680
rtc
socks
sockets
pympler
pylops
